[tests] Skip `deflate_stateful`'s test cases with verification mode enabled for `sw` and `auto` paths (#1248)
The failing tests are the known limitation of the library. There is no quick fix available for the problematic tests, so it was decided to skip them and enable back once the problem in the library is fixed. this will help to achieve green pre-commits and reduce the probability of other hidden errors.